Fixed Ticker.Time should be an int64

+++ market/ticker.go
@@ -2,7 +2,7 @@ package market
 
 // Ticker - Latest price data for an asset
 type Ticker struct {
-	Time   float64
+	Time   int64
 	Candle *Candle
 	Ask    *OrderBookRow
 	Bid    *OrderBookRow
